Refactor: 예외 메시지 코드 추가 및 유효성 검증 로직 개선
- 예외 메시지에 코드 추가하여 가독성 향상
- 유효성 검증 실패 시 메시지 처리 로직 개선 및 글로벌 예외 처리 업데이트
- 쿠폰 관련 비즈니스 로직 및 DTO 에 검증 어노테이션 추가
- 쿠폰 수정 기능 추가 및 서비스 로직 개선

+++ src/main/java/study/back/service/implement/CouponServiceImpl.java
@@ -3,9 +3,11 @@ package study.back.service.implement;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import study.back.dto.request.ModifyCouponRequestDto;
 import study.back.entity.CouponEntity;
+import study.back.exception.BadRequest.InvalidCouponDiscountPercentException;
 import study.back.exception.BadRequest.InvalidCouponNameException;
-import study.back.exception.errors.InternalServerErrorException;
+import study.back.exception.NotFound.CouponNotFoundException;
 import study.back.repository.CouponRepository;
 import study.back.user.entity.UserEntity;
 import study.back.exception.NotFound.UserNotFoundException;
@@ -15,6 +17,7 @@ import study.back.repository.resultSet.UserCouponView;
 import study.back.service.CouponService;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -37,10 +40,6 @@ public class CouponServiceImpl implements CouponService {
     // 새로운 쿠폰 등록하기
     @Override
     public CouponEntity registerCoupon(String name, Integer discountPercent) {
-        if(name == null || name.isBlank()) {
-            throw new InvalidCouponNameException();
-        }
-
         CouponEntity newCoupon = CouponEntity.builder()
                 .name(name)
                 .discountPercent(discountPercent)
@@ -52,9 +51,43 @@ public class CouponServiceImpl implements CouponService {
     // 쿠폰 삭제하기
     @Override
     public void deleteCoupon(Long couponId) {
-        Boolean isDeleteSuccess = couponRepository.deleteCoupon(couponId);
-        if(!isDeleteSuccess) {
-            throw new InternalServerErrorException("쿠폰 삭제 실패");
+        // 쿠폰 여부 검증
+        Boolean isCouponExist = repository.existsCoupon(couponId);
+        if(!isCouponExist) {
+            throw new CouponNotFoundException();
         }
+        // 쿠폰 삭제
+        couponRepository.deleteCoupon(couponId);
+    }
+
+    // 쿠폰 수정하기
+    @Override
+    public void modifyCoupon(ModifyCouponRequestDto modifyCouponRequestDto) {
+        Long couponId = modifyCouponRequestDto.getCouponId();
+        String name = modifyCouponRequestDto.getName();
+        Integer discountPercent = modifyCouponRequestDto.getDiscountPercent();
+
+        // 쿠폰 여부 검증
+        CouponEntity coupon = repository.findCoupon(couponId).orElseThrow(CouponNotFoundException::new);
+
+        // 이름 수정
+        if(name != null && !coupon.getName().equals(name)) {
+            // 이름 검증
+            if(name.isBlank()) {
+                throw new InvalidCouponNameException();
+            }
+            coupon.modifyName(name);
+        }
+
+        // 할인율 수정
+        if(discountPercent != null && !coupon.getDiscountPercent().equals(discountPercent)) {
+            // 할인율 검증
+            if(discountPercent < 1 || discountPercent > 100) {
+                throw new InvalidCouponDiscountPercentException();
+            }
+            coupon.modifyDiscountPercent(discountPercent);
+        }
+
+        // 더티체킹으로 따로 save 하지 않아도 변경 사항 반영됨
     }
 }
